# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PartyPlan API",
    "description": "API de PartyPlan. Événements privés par défaut (RG-EVT-01).",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.partyplan.maxencecoeur.fr",
      "description": "Production"
    }
  ],
  "paths": {
    "/v1/events/{eventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Synthèse d'un événement dont l'appelant est membre.",
        "operationId": "GetEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventSummary": {
        "required": [
          "id",
          "name",
          "description",
          "startsAt",
          "endsAt",
          "address",
          "coverImageUrl",
          "memberCount",
          "presentCount",
          "maybeCount",
          "joinEnabled"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "address": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "memberCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "presentCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maybeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "joinEnabled": {
            "type": "boolean"
          }
        },
        "description": "Vue de synthèse d'un événement, telle que renvoyée par l'API. Type distinct de\nl'entité : le contrat public ne doit pas suivre mécaniquement le schéma de la base."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Events"
    }
  ]
}